+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 -   Use `2022` for `parserOptions.ecmaVersion`
 -   Set Node version to minimum 12.22
+-   Move Unicorn DOM rules to browser configuration
 
 ## [11.0.0][] - 2021-12-28